Fixes edge splitting in BRepAlgo_Loop::Build()
Fixes #1

This affects BRepOffset_MakeOffset::MakeThickSolid() with arc corner.
BRepAlgo_Loop is used by BRepOffset_MakeOffset to build new bounding
wire of offset faces that intersects. It receives a list of edges with
intersecting vertices that are populated by EdgeInter() function in
BRepOffset_Inter2d.cxx.

It seems that BRepAlgo_Loop::CutEdge() relies on the vertex orientation to
decide which segment to keep and which to discard. This logic fails with
shapes in issue #1, which produces wrong segments. This patch keeps
all segments, and relies on the loop finding logic to decide which
segment to keep/discard. The rational being that those edges that are
meant to be discarded can only form loop with some edge(s) that touches
one and only one face. So the fact that only intersecting edges are
passed to BRepAlgo_Loop means those edges will be correctly discarded
because they cannot form any loop.